+++ tests/codeigniter/libraries/Form_validation_test.php
@@ -579,20 +579,6 @@ class Form_validation_test extends CI_TestCase {
 		$this->assertFalse($this->form_validation->regex_match('bar', $regex));
 	}
 
-	public function test_prep_for_form()
-	{
-		$this->form_validation->reset_validation();
-		$error_msg_unprepped = '<error =\'foobar\'">';
-		$error_msg_prepped = '&lt;error =&#39;foobar&#39;&quot;&gt;';
-		$this->form_validation->set_rules('foo', 'label', 'required', array('required' => $error_msg_unprepped));
-		$_POST = array('foo' => '');
-		$this->form_validation->run();
-		$error_arr = $this->form_validation->error_array();
-
-		$this->assertEquals('', $this->form_validation->prep_for_form(''));
-		$this->assertEquals(array('foo' => $error_msg_prepped), $this->form_validation->prep_for_form($error_arr));
-	}
-
 	public function test_prep_url()
 	{
 		$this->assertEquals('', $this->form_validation->prep_url(''));